+++ commons/ava_testsuite/ava_test_suite.go
@@ -11,6 +11,7 @@ func (a AvaTestSuite) GetTests() map[string]testsuite.Test {
 
 	result["fiveStakingNodeGetValidatorsTest"] = FiveNodeStakingNetworkGetValidatorsTest{}
 	result["fiveStakingNodeFullyConnectedTest"] = FiveNodeStakingNetworkFullyConnectedTest{}
+	result["fiveStakingNodeDuplicateNodeIdTest"] = FiveNodeStakingNetworkDuplicateIdTest{}
 	result["stakingNodeRpcWorkflowTest"] = StakingNetworkRpcWorkflowTest{}
 
 	return result