+++ commons/ava_testsuite/staking_gecko_network_tests.go
@@ -25,10 +25,15 @@ const (
 	STAKE_AMOUNT              = int64(30000000000000)
 	DELEGATOR_AMOUNT              = int64(30000000000000)
 	NODE_SERVICE_ID           = 0
-	NODE_CONFIG_ID            = 0
 	DELEGATOR_NODE_SERVICE_ID = 1
+
+	NORMAL_NODE_CONFIG_ID = 0
+
+	// The configuration ID of a service where all servies made with this configuration will have the same cert
+	SAME_CERT_CONFIG_ID = 1
 )
 
+// ================ RPC Workflow Test ===================================
 type StakingNetworkRpcWorkflowTest struct{}
 func (test StakingNetworkRpcWorkflowTest) Run(network interface{}, context testsuite.TestContext) {
 	castedNetwork := network.(ava_networks.TestGeckoNetwork)
@@ -40,11 +45,11 @@ func (test StakingNetworkRpcWorkflowTest) Run(network interface{}, context tests
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get delegator client"))
 	}
-	stakerNodeId, err := stakerClient.AdminApi().GetNodeId()
+	stakerNodeId, err := stakerClient.InfoApi().GetNodeId()
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get staker node ID."))
 	}
-	delegatorNodeId, err := stakerClient.AdminApi().GetNodeId()
+	delegatorNodeId, err := stakerClient.InfoApi().GetNodeId()
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get delegator node ID."))
 	}
@@ -82,7 +87,9 @@ func (test StakingNetworkRpcWorkflowTest) Run(network interface{}, context tests
 		context.Fatal(stacktrace.Propagate(err, "Could not get current stakers."))
 	}
 	logrus.Debugf("Number of current stakers: %d", len(currentStakers))
-	context.AssertTrue(len(currentStakers) == 6)
+	actualNumStakers := len(currentStakers)
+	expectedNumStakers := 6
+	context.AssertTrue(actualNumStakers == expectedNumStakers, stacktrace.NewError("Actual number of stakers, %v, != expected number of stakers, %v", actualNumStakers, expectedNumStakers))
 	// Adding delegators
 	err = highLevelDelegatorClient.AddDelegatorOnSubnet(stakerNodeId, delegatorPchainAddress, DELEGATOR_AMOUNT)
 	if err != nil {
@@ -102,22 +109,28 @@ func (test StakingNetworkRpcWorkflowTest) Run(network interface{}, context tests
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Failed to get account info for account %v.", stakerXchainAddress))
 	}
-	context.AssertTrue(strconv.FormatInt(remainingStakerAva, 10) == xchainAccountInfo.Balance)
+	actualRemainingAva := xchainAccountInfo.Balance
+	expectedRemainingAva := strconv.FormatInt(remainingStakerAva, 10)
+	context.AssertTrue(actualRemainingAva == expectedRemainingAva, stacktrace.NewError("Actual remaining Ava, %v, != expected remaining Ava, %v", actualRemainingAva, expectedRemainingAva))
 }
 func (test StakingNetworkRpcWorkflowTest) GetNetworkLoader() (testsuite.TestNetworkLoader, error) {
-	return getFiveNodeStakingLoader()
+	return getStakingNetworkLoader(map[int]int{
+		NODE_SERVICE_ID:           NORMAL_NODE_CONFIG_ID,
+		DELEGATOR_NODE_SERVICE_ID: NORMAL_NODE_CONFIG_ID,
+	})
 }
 func (test StakingNetworkRpcWorkflowTest) GetTimeout() time.Duration {
 	return 90 * time.Second
 }
 
 
+// =================== Fully Connected Test ==============================
 type FiveNodeStakingNetworkFullyConnectedTest struct{}
 func (test FiveNodeStakingNetworkFullyConnectedTest) Run(network interface{}, context testsuite.TestContext) {
 	castedNetwork := network.(ava_networks.TestGeckoNetwork)
 
-	bootServiceIds := castedNetwork.GetAllBootServiceIds()
-	allServiceIds := append(bootServiceIds, NODE_SERVICE_ID)
+	allServiceIds := castedNetwork.GetAllBootServiceIds()
+	allServiceIds[NODE_SERVICE_ID] = true
 
 	extraStakerClient, err := castedNetwork.GetGeckoClient(NODE_SERVICE_ID)
 	if err != nil {
@@ -134,12 +147,12 @@ func (test FiveNodeStakingNetworkFullyConnectedTest) Run(network interface{}, co
 
 	// collect set of IDs in network
 	nodeIdSet := map[string]bool{}
-	for _, serviceId := range allServiceIds {
+	for serviceId, _ := range allServiceIds {
 		client, err := castedNetwork.GetGeckoClient(serviceId)
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Could not get client for service with ID %v", serviceId))
 		}
-		id, err := client.AdminApi().GetNodeId()
+		id, err := client.InfoApi().GetNodeId()
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Could not get node ID of service with ID %v", serviceId))
 		}
@@ -149,13 +162,13 @@ func (test FiveNodeStakingNetworkFullyConnectedTest) Run(network interface{}, co
 	logrus.Debugf("Network ID Set: %+v", nodeIdSet)
 
 	// verify bootstrapper peer lists have full set of IDs in network.
-	for _, serviceId := range allServiceIds {
+	for serviceId, _ := range allServiceIds {
 		client, err := castedNetwork.GetGeckoClient(serviceId)
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Could not get client for service with ID %v", serviceId))
 		}
-		peers, err := client.AdminApi().GetPeers()
-		nodeId, err := client.AdminApi().GetNodeId()
+		peers, err := client.InfoApi().GetPeers()
+		nodeId, err := client.InfoApi().GetNodeId()
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Could not get peers of service with ID %v", serviceId))
 		}
@@ -168,20 +181,21 @@ func (test FiveNodeStakingNetworkFullyConnectedTest) Run(network interface{}, co
 		for expectedNodeId, _ := range nodeIdSet {
 			// Nodes are expected to have all other nodes in the network on their peer list.
 			if nodeId != expectedNodeId {
-				context.AssertTrue(peerSet[expectedNodeId])
+				context.AssertTrue(peerSet[expectedNodeId], stacktrace.NewError("Didn't find node ID %v in peer set", expectedNodeId))
 			}
 		}
 	}
 }
 
 func (test FiveNodeStakingNetworkFullyConnectedTest) GetNetworkLoader() (testsuite.TestNetworkLoader, error) {
-	return getFiveNodeStakingLoader()
+	return getStakingNetworkLoader(map[int]int{
+		NODE_SERVICE_ID:           NORMAL_NODE_CONFIG_ID,
+	})
 }
 
 func (test FiveNodeStakingNetworkFullyConnectedTest) GetTimeout() time.Duration {
 	return 60 * time.Second
 }
-
 
 // =============== Get Validators Test ==================================
 type FiveNodeStakingNetworkGetValidatorsTest struct{}
@@ -214,33 +228,282 @@ func (test FiveNodeStakingNetworkGetValidatorsTest) Run(network interface{}, con
 		logrus.Infof("Validator ID: %s", validator.Id)
 	}
 	// TODO change this to be specific
-	context.AssertTrue(len(validators) >= 1)
+	context.AssertTrue(len(validators) >= 1, stacktrace.NewError("No validators found"))
 }
 
 func (test FiveNodeStakingNetworkGetValidatorsTest) GetNetworkLoader() (testsuite.TestNetworkLoader, error) {
-	return getFiveNodeStakingLoader()
+	return getStakingNetworkLoader(map[int]int{
+		NODE_SERVICE_ID:           NORMAL_NODE_CONFIG_ID,
+	})
 }
 
 func (test FiveNodeStakingNetworkGetValidatorsTest) GetTimeout() time.Duration {
 	return 30 * time.Second
 }
 
+// =============== Duplicate Node ID Test ==============================
+type FiveNodeStakingNetworkDuplicateIdTest struct {}
+func (f FiveNodeStakingNetworkDuplicateIdTest) Run(network interface{}, context testsuite.TestContext) {
+	castedNetwork := network.(ava_networks.TestGeckoNetwork)
+
+	bootServiceIds := castedNetwork.GetAllBootServiceIds()
+
+	allServiceIds := make(map[int]bool)
+	allServiceIds[NODE_SERVICE_ID] = true
+	for bootServiceId, _ := range bootServiceIds {
+		allServiceIds[bootServiceId] = true
+	}
+
+	allGeckoClients := make(map[int]*gecko_client.GeckoClient)
+	allNodeIds := make(map[int]string)
+	for serviceId, _ := range allServiceIds {
+		client, err := castedNetwork.GetGeckoClient(serviceId)
+		if err != nil {
+			context.Fatal(stacktrace.Propagate(err, "An error occurred getting the Gecko client for service with ID %v", serviceId))
+		}
+		allGeckoClients[serviceId] = client
+		nodeId, err := client.InfoApi().GetNodeId()
+		if err != nil {
+			context.Fatal(stacktrace.Propagate(err, "An error occurred getting the Gecko node ID for service with ID %v", serviceId))
+		}
+		allNodeIds[serviceId] = nodeId
+	}
+
+	logrus.Info("Verifying that initial network state is as expected...")
+	for serviceId, _ := range allServiceIds {
+		acceptableNodeIds := make(map[string]bool)
+		for iterServiceId, nodeId := range allNodeIds {
+			if serviceId != iterServiceId {
+				acceptableNodeIds[nodeId] = true
+			}
+		}
+		verifyExpectedPeers(context, serviceId, allGeckoClients[serviceId], acceptableNodeIds, len(allServiceIds) - 1, false)
+	}
+
+	// We'll need these later
+	originalServiceIds := make(map[int]bool)
+	for serviceId, _ := range allServiceIds {
+		originalServiceIds[serviceId] = true
+	}
+
+	logrus.Debugf("Service IDs before adding any nodes: %v", allServiceIds)
+	logrus.Debugf("Gecko node IDs before adding any nodes: %v", allNodeIds)
+
+	// Add the first dupe node ID (should look normal from a network perspective
+	badServiceId1 := 1
+	logrus.Info("Adding first node with soon-to-be-duplicated node ID...")
+	checker1, err := castedNetwork.AddService(SAME_CERT_CONFIG_ID, badServiceId1)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to create first dupe node ID service with ID %v", badServiceId1))
+	}
+	if err := checker1.WaitForStartup(); err != nil {
+		context.Fatal(stacktrace.Propagate(err, "An error occurred waiting for first dupe node ID service with ID %v to start", badServiceId1))
+	}
+	allServiceIds[badServiceId1] = true
+
+	badServiceClient1, err := castedNetwork.GetGeckoClient(badServiceId1)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "An error occurred getting the Gecko client for the first dupe node ID service with ID %v", badServiceId1))
+	}
+	allGeckoClients[badServiceId1] = badServiceClient1
+
+	badServiceNodeId1, err := badServiceClient1.InfoApi().GetNodeId()
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get node ID from first dupe node ID service with ID %v", badServiceId1))
+	}
+	allNodeIds[badServiceId1] = badServiceNodeId1
+
+	logrus.Info("Successfully added first node with soon-to-be-duplicated ID")
+
+	// Verify that the new node got accepted by everyone
+	logrus.Infof("Verifying that the new node with service ID %v was accepted by all bootstrappers...", badServiceId1)
+	for serviceId, _ := range allServiceIds {
+		acceptableNodeIds := make(map[string]bool)
+
+		// All original nodes should have the boot nodes (though a boot node won't have itself)
+		for bootServiceId, _ := range bootServiceIds {
+			if serviceId != bootServiceId {
+				bootNodeId := allNodeIds[bootServiceId]
+				acceptableNodeIds[bootNodeId] = true
+			}
+		}
+
+		// Boot nodes will also have the other two nodes
+		if _, found := bootServiceIds[serviceId]; found {
+			acceptableNodeIds[allNodeIds[NODE_SERVICE_ID]] = true
+			acceptableNodeIds[badServiceNodeId1] = true
+			verifyExpectedPeers(context, serviceId, allGeckoClients[serviceId], acceptableNodeIds, len(allServiceIds) - 1, false)
+		} else {
+			verifyExpectedPeers(context, serviceId, allGeckoClients[serviceId], acceptableNodeIds, len(bootServiceIds), false)
+		}
+	}
+	logrus.Infof("New node with service ID %v was accepted by all bootstrappers", badServiceId1)
+
+	// Now, add a second node with the same ID
+	badServiceId2 := 2
+	logrus.Infof("Adding second node with service ID %v which will be a duplicated node ID...", badServiceId2)
+	checker2, err := castedNetwork.AddService(SAME_CERT_CONFIG_ID, badServiceId2)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to create second dupe node ID service with ID %v", badServiceId2))
+	}
+	if err := checker2.WaitForStartup(); err != nil {
+		context.Fatal(stacktrace.Propagate(err, "An error occurred waiting for second dupe node ID service to start"))
+	}
+	allServiceIds[badServiceId2] = true
+
+	badServiceClient2, err := castedNetwork.GetGeckoClient(badServiceId2)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "An error occurred getting the Gecko client for the second dupe node ID service with ID %v", badServiceId2))
+	}
+	allGeckoClients[badServiceId2] = badServiceClient2
+
+	badServiceNodeId2, err := badServiceClient2.InfoApi().GetNodeId()
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get node ID from first dupe node ID service with ID %v", badServiceId2))
+	}
+	allNodeIds[badServiceId2] = badServiceNodeId2
+	logrus.Info("Second node added, causing duplicate node ID")
+
+	// At this point, it's undefined what happens with the two nodes with duplicate IDs; verify that the original nodes
+	//  in the network operate normally amongst themselves
+	logrus.Info("Connection behaviour to nodes with duplicate IDs is undefined, so verifying that the original nodes connect as expected...")
+	for serviceId, _ := range originalServiceIds {
+		acceptableNodeIds := make(map[string]bool)
+
+		// All original nodes should have the boot nodes (though a boot node won't have itself)
+		for bootServiceId, _ := range bootServiceIds {
+			if serviceId != bootServiceId {
+				bootNodeId := allNodeIds[bootServiceId]
+				acceptableNodeIds[bootNodeId] = true
+			}
+		}
+
+		if _, found := bootServiceIds[serviceId]; found {
+			// Boot nodes should have the original node, one of the duplicates, and MAY have the duplicate nodes
+			acceptableNodeIds[allNodeIds[NODE_SERVICE_ID]] = true
+			acceptableNodeIds[badServiceNodeId1] = true
+			acceptableNodeIds[badServiceNodeId2] = true
+			verifyExpectedPeers(context, serviceId, allGeckoClients[serviceId], acceptableNodeIds, len(originalServiceIds) - 1, true)
+		} else {
+			// The original non-boot node should have exactly the boot nodes
+			verifyExpectedPeers(context, serviceId, allGeckoClients[serviceId], acceptableNodeIds, len(bootServiceIds), false)
+		}
+	}
+	logrus.Info("Verified that original nodes are still connected to each other")
+
+	// Now, kill the first dupe node to leave only the second (who everyone should connect with)
+	logrus.Info("Removing first node with duplicate ID...")
+	if err := castedNetwork.RemoveService(badServiceId1); err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not remove the first service with duped node ID"))
+	}
+	delete(allServiceIds, badServiceId1)
+	delete(allGeckoClients, badServiceId1)
+	delete(allNodeIds, badServiceId1)
+	logrus.Info("Successfully removed first node with duplicate ID, leaving only the second")
+
+	// Now that the first duped node is gone, verify that the original node is still connected to just boot nodes and
+	//  the second duped-ID node is now accepted by the boot nodes
+	logrus.Info("Verifying that the network has connected to the second node with a previously-duplicated node ID...")
+	for serviceId, _ := range allServiceIds {
+		acceptableNodeIds := make(map[string]bool)
+
+		// All nodes should have the boot nodes (though a boot node won't have itself)
+		for bootServiceId, _ := range bootServiceIds {
+			if serviceId != bootServiceId {
+				bootNodeId := allNodeIds[bootServiceId]
+				acceptableNodeIds[bootNodeId] = true
+			}
+		}
+
+		// Boot nodes should have all nodes
+		if _, found := bootServiceIds[serviceId]; found {
+			acceptableNodeIds[allNodeIds[NODE_SERVICE_ID]] = true
+			acceptableNodeIds[badServiceNodeId2] = true
+			verifyExpectedPeers(context, serviceId, allGeckoClients[serviceId], acceptableNodeIds, len(allServiceIds) - 1, false)
+		} else {
+			verifyExpectedPeers(context, serviceId, allGeckoClients[serviceId], acceptableNodeIds, len(bootServiceIds), false)
+		}
+	}
+	logrus.Info("Verified that the network has settled on the second node with previously-duplicated ID")
+}
+
+func (f FiveNodeStakingNetworkDuplicateIdTest) GetNetworkLoader() (testsuite.TestNetworkLoader, error) {
+	return getStakingNetworkLoader(map[int]int{
+		NODE_SERVICE_ID:           NORMAL_NODE_CONFIG_ID,
+	})
+}
+
+func (f FiveNodeStakingNetworkDuplicateIdTest) GetTimeout() time.Duration {
+	return 120 * time.Second
+}
+
 // =============== Helper functions =============================
 
-// TODO TODO TODO Rename this
-func getFiveNodeStakingLoader() (testsuite.TestNetworkLoader, error) {
+/*
+Args:
+	desiredServices: Mapping of service_id -> configuration_id for all services *in addition to the boot nodes* that the user wants
+ */
+func getStakingNetworkLoader(desiredServices map[int]int) (testsuite.TestNetworkLoader, error) {
 	serviceConfigs := map[int]ava_networks.TestGeckoNetworkServiceConfig{
-		NODE_CONFIG_ID:           *ava_networks.NewTestGeckoNetworkServiceConfig(true, ava_services.LOG_LEVEL_DEBUG),
+		NORMAL_NODE_CONFIG_ID: *ava_networks.NewTestGeckoNetworkServiceConfig(true, ava_services.LOG_LEVEL_DEBUG),
+		SAME_CERT_CONFIG_ID:   *ava_networks.NewTestGeckoNetworkServiceConfig(false, ava_services.LOG_LEVEL_DEBUG),
 	}
 	return ava_networks.NewTestGeckoNetworkLoader(
 		ava_services.LOG_LEVEL_DEBUG,
 		true,
 		serviceConfigs,
-		map[int]int{
-			NODE_SERVICE_ID:           NODE_CONFIG_ID,
-			DELEGATOR_NODE_SERVICE_ID: NODE_CONFIG_ID,
-		},
+		desiredServices,
 		2,
 		2)
 }
 
+/*
+Verifies that a node's actual peers are what we expect
+
+Args:
+	context: Test context (used for failing if there's a problem)
+	serviceId: Service ID of the node whose peers are being examined
+	client: Gecko client for the node being examined
+	acceptableNodeIds: A "set" of acceptable node IDs where, if a peer doesn't have this ID, the test will be failed
+	expectedNumPeers: The number of peers we expect this node to have
+	atLeast: If true, indicates that the number of peers must be AT LEAST the expected number of peers; if false, must be exact
+ */
+func verifyExpectedPeers(
+			context testsuite.TestContext,
+			serviceId int,
+			client *gecko_client.GeckoClient,
+			acceptableNodeIds map[string]bool,
+			expectedNumPeers int,
+			atLeast bool) {
+	peers, err := client.InfoApi().GetPeers()
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to get peers from service with ID %v", serviceId))
+	}
+
+	actualNumPeers := len(peers)
+	var condition bool
+	var operatorAsserted string
+	if atLeast {
+		condition = actualNumPeers >= expectedNumPeers
+		operatorAsserted = ">="
+	} else {
+		condition = actualNumPeers == expectedNumPeers
+		operatorAsserted = "=="
+	}
+	context.AssertTrue(
+		condition,
+		stacktrace.NewError(
+			"Service ID %v actual num peers, %v, is not %v expected num peers, %v",
+			serviceId,
+			actualNumPeers,
+			operatorAsserted,
+			expectedNumPeers,
+		),
+	)
+
+	// Verify that IDs of the peers we have are in our list of acceptable IDs
+	for _, peer := range peers {
+		_, found := acceptableNodeIds[peer.Id]
+		context.AssertTrue(found, stacktrace.NewError("Service ID %v has a peer with node ID %v that we don't recognize", serviceId, peer.Id))
+	}
+}