+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# TBD
+
+# 0.3.1
+* Specify --http-host CLI flag in GetStartCommand to have RPC calls bind to publicIP
+* Migrate AdminAPI endpoints to the new InfoAPI
+* Parse error message on XChain Imports in order to wait for PChain transaction to be accepted
+* Added tests for XChain endpoints
+* Remove `FiveNodeStakingNetworkBasicTest` (wasn't being used)
+* Test if the network functions as expected when nodes with duplicate node IDs occur
+* Drop default loglevel for initializer & controller down to DEBUG
+
 # 0.3.0
 * Removed startPortRange and endPortRange CLI args
 * Declare timeouts for tests